ASoC: cs35l41: Add PAUSE mbox command
DSP is paused and resumed on AMP power event if a FW is
loaded. Even if PCM Source is not connected to DSP,
it will be resumed on AMP power on event. Setting
'MBOX CSPL Command' to CSPL_MBOX_CMD_PAUSE will allow
to keep DSP paused after AMP enable event.

Signed-off-by: Vlad Karpovich <Vlad.Karpovich@cirrus.com>