ASoC: wm_adsp: Associate buffer regions with DSP rather than firmware
Firmwares have always used the same layout of buffer regions, simply
specifying zero size for unused regions. However, in the future
different types of DSP will be added with new memory types which will
require a different layout of buffer regions. As such it makes more
sense to move the specification of the buffer regions from being tided
to the firmware to the DSP itself.

Signed-off-by: Andrew Ford <aford@opensource.cirrus.com>
Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>