ASoC: cs35l41: VBSTMON DAPM Switch
Add a DAPM switch that will create a route
from VBSTMON to the main amp output.

Unlike the other monitoring data sources,
VBSTMON is valid without audio inputs.
To allow for VBSTMON to be monitored without
opening the Playback path, this patch adds
a switch that can be enabled to power on the
other relevant widgets for checking VBSTMON.

Signed-off-by: David Rhodes <david.rhodes@cirrus.com>