ASoC: cs35l41: Change headers path to relative
The relative path does not work in Android build tree
because the c code is moved to the out folder for build
but the headers are not moved. The headers path shound be
pointing to the original module location, not to the path inside
the out folder.
Since AndroidKernelModiles.mk pass source locatin with kernel offset
prefix, it should be stripped before constracting the obsolute path.
For example, if kernel in Anroid tree located at kernel/msm-4.19/
AndroidKernelModiles.mk will add ../../ to the $(src).

Change-Id: I9f984a1edaa02ba9a8272b069cd3fa38bf2eb8a8
Signed-off-by: Vlad Karpovich <Vlad.Karpovich@cirrus.com>